refactor: 🛠️ make contentful-infra an implicit lib
