--- libs/contentful-infra/README.md
@@ -1,7 +1,0 @@
-# contentful-api
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test contentful-api` to execute the unit tests.
