--- libs/contentful-infra/project.json
@@ -1,21 +1,0 @@
-{
-  "name": "contentful-infra",
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "projectType": "library",
-  "sourceRoot": "libs/contentful-api/src",
-  "prefix": "mc",
-  "tags": ["scope:contentful", "type:infra", "name:contentful-infra"],
-  "generators": {},
-  "targets": {
-    "lint": {
-      "executor": "@nx/eslint:lint"
-    },
-    "test": {
-      "executor": "@nx/jest:jest",
-      "options": {
-        "jestConfig": "{projectRoot}/jest.config.ts"
-      },
-      "outputs": ["{workspaceRoot}/coverage/libs/contentful-api"]
-    }
-  }
-}
