+++ libs/contentful/infra/lib/schema.graphql
