+++ libs/contentful/infra/lib/schema.ts
